Cart Function part 4, find the product id

+++ public/cart.php
@@ -47,31 +47,42 @@
 
         foreach ($_SESSION as $name => $value) {
             
-            if(substr($name, 0, 8) == 'product_'){
+            if($value > 0){
 
-                $query = query("SELECT * FROM products");
-                confirm($query);
+                if(substr($name, 0, 8) == 'product_'){
 
-                while($row = fetchArray($query)){
+                    $length = strlen($name);
+                    $len = $length - 8;
+                    $id = substr($name, 8, $len);
+                    // $length = explode('_', $name); 
+                    // $id = substr($name, 8, $length[1]);
+                    
 
-                    $product = <<<DELIMETER
-
-                            <tr>
-                                <td>{$row['product_title']}</td>
-                                <td>$23</td>
-                                <td>3</td>
-                                <td>2</td>
-                                <td>
-                                    <a class="btn btn-warning" href="cart.php?remove={$row['product_id']}"><span class="glyphicon glyphicon-minus"></span></a>
-                                    <a class="btn btn-success" href="cart.php?add={$row['product_id']}"><span class="glyphicon glyphicon-plus"></span></a>
-                                    <a class="btn btn-danger" href="cart.php?delete={$row['product_id']}"><span class="glyphicon glyphicon-remove"></span></a>
-                                </td>
-                            
-                            </tr>
-
-                    DELIMETER;
-
-                    echo $product;
+                    $query = query("SELECT * FROM products WHERE product_id = " . escapeString($id) . " " );
+                    confirm($query);
+    
+                    while($row = fetchArray($query)){
+    
+                        $product = <<<DELIMETER
+    
+                                <tr>
+                                    <td>{$row['product_title']}</td>
+                                    <td>$23</td>
+                                    <td>3</td>
+                                    <td>2</td>
+                                    <td>
+                                        <a class="btn btn-warning" href="cart.php?remove={$row['product_id']}"><span class="glyphicon glyphicon-minus"></span></a>
+                                        <a class="btn btn-success" href="cart.php?add={$row['product_id']}"><span class="glyphicon glyphicon-plus"></span></a>
+                                        <a class="btn btn-danger" href="cart.php?delete={$row['product_id']}"><span class="glyphicon glyphicon-remove"></span></a>
+                                    </td>
+                                
+                                </tr>
+    
+                        DELIMETER;
+    
+                        echo $product;
+                    }
+    
                 }
 
             }
